+++ src/contexts/CoreContext/domain/interfaces/repositories/IFreelancerRepository.ts
@@ -2,4 +2,4 @@ import { IRepository } from "@/contexts/Shared/domain/repository/IRepository";
 import { Freelancer } from "../../entities/Freelancer";
 
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface IFreelancerRepository extends IRepository<Freelancer> {}
+export interface IFreelancerRepository extends IRepository<Freelancer> { }
