+++ src/contexts/CoreContext/mappers/FreelancerMapper.ts
@@ -1,9 +1,7 @@
-import { Freelancer } from "../domain/entities/Freelancer";
 import { Freelancer as PrismaFreelancer } from "@/generated/prisma";
 import { UniqueEntityID } from "@/contexts/Shared/domain/UniqueEntityID";
 import { About } from "../domain/valueObjects/About";
 import { UserId } from "../domain/valueObjects/UserId";
-import { IFreelancerProfileDto } from "../domain/interfaces/dtos/IFreelancerProfileDto";
 
 export default class FreelancerMapper {
   static persistanceToDomain(
